# [Discussion] Using Rapscallion with NextJS