--- a/examples/express-with-rapscallion/.babelrc
+++ b/examples/express-with-rapscallion/.babelrc
@@ -1,0 +1,19 @@
+{
+	"env": {
+		"development": {
+			"presets": "next/babel"
+		},
+		"production": {
+			"presets": "next/babel"
+		},
+		"test": {
+			"presets": [
+				["env", {
+					"modules": "commonjs"
+				}],
+				"next/babel",
+				"rapscallion/babel-plugin-server"
+			]
+		}
+	}
+}