--- a/examples/express-with-rapscallion/README.md
+++ b/examples/express-with-rapscallion/README.md
@@ -1,0 +1,1 @@
+# Express with Rapscallion