--- a/examples/express-with-rapscallion/pages/index.js
+++ b/examples/express-with-rapscallion/pages/index.js
@@ -1,0 +1,8 @@
+import React from 'react'
+
+export default () => (
+  <ul>
+    <li><a href='/b'>Default Next.JS rendering</a></li>
+    <li><a href='/a'>Rapscallion rendering, with component caching</a></li>
+  </ul>
+)