@arunoda 2017-06-16
This is pretty great.
I'd like to work on these as soon as we ship 3.0