@danoc 2017-08-07
Does React 16's improvements to SSR change the approach here?

* Proposal: ReactDOMServer render to Stream – https://github.com/facebook/react/issues/6420
* React 16 beta – https://github.com/facebook/react/issues/10294